requirements.txt
aiohttp==3.6.2
alembic==1.8.1
anyio==3.6.2
api-display-purposes==0.0.3
appdirs==1.4.4
appnope==0.1.0
APScheduler==3.6.3
argon2-cffi==20.1.0
astroid==2.4.2
asttokens==2.2.1
async-timeout==3.0.1
attrs==20.2.0
Babel==2.8.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.1
bleach==3.1.5
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
cffi==1.14.2
chardet==3.0.4
charset-normalizer==2.1.1
clarifai==2.6.2
click==8.1.3
comm==0.1.2
configparser==3.8.1
cvxopt==1.2.5
cvxpy==1.1.5
cycler==0.10.0
db-dtypes==1.0.4
debugpy==1.6.6
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
dnspython==2.2.1
dominate==2.7.0
EasyProcess==1.1
ecos==2.0.7.post1
email-validator==1.3.0
emoji==1.6.3
entrypoints==0.3
executing==1.2.0
filelock==3.0.12
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-Datepicker==0.14
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Moment==1.0.5
Flask-PageDown==0.4.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
future==0.18.2
google-api-core==1.34.0
google-api-python-client==1.12.3
google-auth==2.14.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.7.1
google-cloud-bigquery==3.3.6
google-cloud-bigquery-storage==2.16.2
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
greenlet==2.0.1
grpcio==1.50.0
grpcio-status==1.48.2
gspread==5.6.2
h11==0.14.0
httpcore==0.16.3
httplib2==0.21.0
httpx==0.23.3
idna==2.10
imageio==2.9.0
importlib-metadata==4.11.3
importlib-resources==5.7.1
instapy==0.6.16
ipykernel==6.21.2
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isort==4.3.21
itsdangerous==2.1.2
jedi==0.17.2
Jinja2==3.1.2
joblib==0.16.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-console==6.6.2
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyterlab-widgets==3.0.5
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.9.1
Mako==1.2.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.3.1
matplotlib-inline==0.1.6
mccabe==0.6.1
MeaningCloud-python==2.0.0
mglearn==0.1.9
mistune==0.8.4
multidict==4.7.6
multitasking==0.0.11
nbconvert==5.6.1
nbformat==5.0.7
nest-asyncio==1.5.6
notebook==6.1.3
npm==0.1.1
numpy==1.23.4
oauth2client==4.1.3
oauthlib==3.2.2
openai==0.26.2
optional-django==0.1.0
osqp==0.6.1
packaging==21.3
pandas==1.5.1
pandas-gbq==0.17.9
pandoc==1.0.2
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pipenv==2020.6.2
platformdirs==3.0.0
plotly==5.11.0
ply==3.11
plyer==2.0.0
prometheus-client==0.8.0
prompt-toolkit==3.0.38
proto-plus==1.22.1
protobuf==3.20.3
psutil==5.9.4
psycopg2==2.9.5
ptyprocess==0.6.0
pure-eval==0.2.2
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydata-google-auth==1.4.0
pyee==7.0.4
Pygments==2.6.1
PyJWT==2.6.0
pylint==2.5.3
PyMySQL==1.0.2
pyparsing==3.0.9
pyportfolioopt==1.2.5
pyppeteer==0.2.2
pyrsistent==0.16.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-telegram-bot==20.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyVirtualDisplay==2.2
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
regex==2022.1.18
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
scikit-learn==0.23.2
scipy==1.5.2
scs==2.1.2
secure-smtplib==0.1.1
selenium==3.141.0
semantic-version==2.8.5
Send2Trash==1.5.0
setuptools-rust==1.1.2
simplejson==3.17.2
six==1.16.0
slack==0.0.2
sniffio==1.3.0
soupsieve==2.0.1
SQLAlchemy==1.4.43
stack-data==0.6.2
tenacity==8.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
twilio==6.45.4
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
uritemplate==3.0.1
urllib3==1.26.12
virtualenv==20.0.25
virtualenv-clone==0.5.4
visitor==0.1.3
wcwidth==0.2.5
webdriverdownloader==1.1.0.3
webencodings==0.5.1
websockets==8.1
Werkzeug==2.2.2
widgetsnbextension==4.0.5
wrapt==1.12.1
WTForms==3.0.1
xlrd==1.2.0
yahoo-finance==1.4.0
yarl==1.6.0
yfinance==0.1.85
zipp==3.8.0
